+++ vite-react/src/components/app/PageError.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import { useErrorBoundary } from 'react-error-boundary';
-import { useNavigate } from '../router';
+import { useNavigate } from '../../router';
 
 interface PageErrorProps {
   errorText: string;
