+++ vite-react/src/services/QueryKey.ts
@@ -1,5 +1,6 @@
 enum QueryKey {
   GET_POST = 'getPost',
+  GET_POSTS = 'getPosts',
 }
 
 export default QueryKey;
