--- vite-react/src/services/queries/getPostsQuery.ts
@@ -1,8 +1,0 @@
-import postsService from '../services/postsService';
-
-const getPostsQuery = () => ({
-  queryKey: ['getPosts'],
-  queryFn: () => postsService.getPosts(),
-});
-
-export default getPostsQuery;
